requirements.txt
torch>=1.12
numpy>=1.21
h5py>=3.6
Pillow>=9.0
